+++ R/read.R
@@ -98,7 +98,7 @@ is_functions = function(x) {
 read_stars = function(.x, ..., options = character(0), driver = character(0),
 		sub = TRUE, quiet = FALSE, NA_value = NA_real_, along = NA_integer_,
 		RasterIO = list(), proxy = is_functions(.x) || (!length(curvilinear) &&
-				is_big(.x, sub = sub, driver=driver, normalize_path = normalize_path, ...)),
+				is_big(.x, sub = sub, driver = driver, normalize_path = normalize_path, ...)),
 		curvilinear = character(0), normalize_path = TRUE, RAT = character(0),
 		tolerance = 1e-10, exclude = "") {
 
@@ -334,92 +334,4 @@ get_data_units = function(data) {
 		else
 			structure(data, units = NULL)
 	}
-}
-
-#' Read data using GDAL's multidimensional array API (experimental)
-#' 
-#' Read data using GDAL's multidimensional array API (experimental)
-#' @param x data source name
-#' @param variable name of the array to be read
-#' @param options array opening options
-#' @param raster names of the raster variables (default: first two)
-#' @param offset integer; offset for each dimension (pixels) of sub-array to read (default: 0,0,0,...) (requires sf >= 1.0-9)
-#' @param count integer; size for each dimension (pixels) of sub-array to read (default: read all) (requires sf >= 1.0-9)
-#' @param step integer; step size for each dimension (pixels) of sub-aray to read (requires sf >= 1.0-9)
-#' @param proxy logical; return proxy object? (not functional yet)
-#' @param debug logical; print debug info?
-#' @details it is assumed that the first two dimensions are easting / northing
-#' @param ... ignored
-#' @export
-read_mdim = function(x, variable = character(0), ..., options = character(0), raster = NULL,
-					 offset = integer(0), count = integer(0), step = integer(0), proxy = FALSE, 
-					 debug = FALSE) {
-
-	# when releasing to CRAN, require sf 1.0-9 and drop second option
-	ret = if (packageVersion("sf") >= "1.0-9")
-			gdal_read_mdim(x, variable, options, rev(offset), rev(count), rev(step), proxy, debug)
-		else
-			gdal_read_mdim(x, variable, options)
-
-#	if (packageVersion("sf") >= "1.0-9") {
-#		message("update stars to > 0.5-6")
-#		return(NULL)
-#	}
-#	ret = gdal_read_mdim(x, variable, options)
-	create_units = function(x) {
-		u <- attr(x, "units")
-		if (is.null(u) || u == "")
-			x
-		else {
-			if (!is.null(a <- attr(x, "attributes")) && !is.na(cal <- a["calendar"]) && 
-						cal %in% c("360_day", "365_day", "noleap"))
-				get_pcict(x, u, cal)
-			else {
-				u = units::set_units(x, u, mode = "standard")
-				p = try(as.POSIXct(u), silent = TRUE)
-				if (inherits(p, "POSIXct"))
-					p
-				else
-					u
-			}
-		}
-	}
-	l = rev(lapply(ret$dimensions, function(x) create_units(x$values[[1]])))
-	if (length(offset) != 0 || length(step) != 0 || length(count) != 0) {
-		if (length(offset) == 0)
-			offset = rep(0, length(l))
-		if (length(step) == 0)
-			step = rep(1, length(l))
-		if (length(count) == 0)
-			count = floor((lengths(l) - offset)/step)
-		for (i in seq_along(l)) {
-			l[[i]] = l[[i]][seq(from = offset[i]+1, length.out = count[i], by = step[i])]
-		}
-	}
-	d = mapply(function(x, i) create_dimension(values = x, is_raster = i %in% 1:2), l, seq_along(l),
-			SIMPLIFY = FALSE)
-	if (is.null(raster))
-		raster = get_raster(dimensions = names(d)[1:2])
-	else
-		raster = get_raster(dimensions = raster)
-
-	if (proxy)
-		stop("proxy not yet implemented in read_mdim()")
-
-	# handle array units:
-	for (i in seq_along(ret$array_list))
-		if (nchar(u <- attr(ret$array_list[[i]], "units")))
-			ret$array_list[[i]] = units::set_units(ret$array_list[[i]], u, mode = "standard")
-	lst = lapply(ret$array_list, function(x) structure(x, dim = rev(dim(x))))
-	st_set_crs(st_stars(lst, dimensions = structure(d, raster = raster, class = "dimensions")),
-		ret$srs)
-}
-
-write_mdim = function(x, filename, ...) {
-	stopifnot(inherits(x, "stars"))
-	to_units = function(x) if (inherits(x, c("POSIXct", "Date"))) units::as_units(x) else x
-	dimension_values = rev(lapply(expand_dimensions(x), to_units))
-	units = sapply(dimension_values, function(x) if(inherits(x, "units")) as.character(units(x)) else "")
-	gdal_write_mdim(st_as_stars(x), filename, dimension_values, units)
-	invisible(x)
 }